% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abm.table}
\alias{abm.table}
\title{EFForTS-ABM table with parameters and metrics}
\format{
EFForTS-ABM table with parameters and metrics
}
\usage{
abm.table
}
\description{
The dataset contains a complete table with parameters and metrics of EFForTS-ABM.
}
\keyword{datasets}
