% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.management.default.R
\name{get.abm.management.default}
\alias{get.abm.management.default}
\title{Get a skeleton for the management parameter file of EFForTS-ABM}
\usage{
get.abm.management.default(crop = NULL, n = 1)
}
\arguments{
\item{crop}{optionally, select a crop type for defaults}

\item{n}{number of management skeletons}
}
\value{
a list with the parameter format of the main parameter file
}
\description{
Get a skeleton for the management parameter file of EFForTS-ABM
}
\details{
Get a skeleton for the management parameter file of EFForTS-ABM
}
\examples{
get.abm.management.default(crop="rubber")

}
