% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.metrics.R
\name{get.abm.metrics}
\alias{get.abm.metrics}
\title{Report all metrics as a vector}
\usage{
get.abm.metrics()
}
\value{
all metrics as a vector
}
\description{
Report all metrics as a vector
}
\details{
Report all metrics as a vector
}
\examples{
get.abm.metrics()

}
