% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.metrics.turtles.R
\name{get.abm.metrics.turtles}
\alias{get.abm.metrics.turtles}
\title{Report turtles metrics as a list}
\usage{
get.abm.metrics.turtles()
}
\value{
turtles metrics as a list
}
\description{
Report turtles metrics as a list
}
\details{
Report turtles metrics as a list
}
\examples{
get.abm.metrics.turtles()

}
