% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.parfiles.R
\name{get.abm.parfiles}
\alias{get.abm.parfiles}
\title{Get parameter files from efforts-abm folder}
\usage{
get.abm.parfiles(abmfolder)
}
\arguments{
\item{abmfolder}{folder destination of EFForTS-ABM model}
}
\value{
list with parameter definitions
}
\description{
Get parameter files from efforts-abm folder
}
\details{
Get parameter files from efforts-abm folder
}
\examples{

\dontrun{
get.abm.parfiles()
}

}
