% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.exec.R
\name{hpc.exec}
\alias{hpc.exec}
\title{Executes shell commands on the remote host (HPC) via ssh}
\usage{
hpc.exec(command = NA, user = NA, host = "transfer.gwdg.de", key = NA)
}
\arguments{
\item{command}{command string}

\item{user}{user name on the remote host}

\item{host}{host server address}

\item{key}{link to the ssh key file on the local system}
}
\description{
Executes shell commands on the remote host (HPC) via ssh
}
\details{
Executes shell commands on the remote host (HPC) via ssh
}
