% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.GBM.cluster.series.R
\name{prices.GBM.cluster.series}
\alias{prices.GBM.cluster.series}
\title{Cluster multiple GBM series}
\usage{
prices.GBM.cluster.series(series, clusterlist)
}
\arguments{
\item{series}{tibble with at least 2 series of GBM price data}

\item{clusterlist}{list with cluster objects, derived from prices.GBM.cluster()}
}
\value{
a tibble with series data and cluster ids
}
\description{
Get parameter files from efforts-abm folder
}
\details{
The function extracts cluster ids from the cluster objects from the prices.GBM.cluster function and joins them to the original series data.
}
