% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.zoo.R
\name{prices.zoo}
\alias{prices.zoo}
\title{Get monthly price data as zoo object from included package data prices.wb}
\usage{
prices.zoo(
  start.date = as.Date("1960-01-01"),
  end.date = as.Date("2020-01-01")
)
}
\arguments{
\item{start.date}{start date of time series in format as.Date("yyyy-mm-dd")}

\item{end.date}{end date of time series as.Date("yyyy-mm-dd")}
}
\value{
a zoo object with price time series
}
\description{
Get parameter files from efforts-abm folder
}
\details{
Loads monthly price data for the selected crop from the included data file prices.wb.
The prices are then converted into an zoo object.
Price data has been stored within the package and was downloaded from World Bank Data on March 3rd, 2020.
Prices for Oilpalm and Rubber were extracted and converted to prices in Dollar per ton from the Pink Sheet data (World Bank).
}
\examples{
\dontrun{
prices.zoo()
}

}
