% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refforts-package.R
\docType{package}
\name{Refforts-package}
\alias{Refforts}
\alias{Refforts-package}
\title{Refforts: an R package for performing analyes with EFForTS tools}
\description{
What the package does (one paragraph).
}
\details{
Refforts: an R package for performing analyes with EFForTS tools
}
\author{
Jan Salecker \email{jsaleck@gwdg.de}
}
\keyword{package}
