% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.nl.variable.R
\name{set.nl.variable}
\alias{set.nl.variable}
\title{Set a variable in the exepriment of the provided nl object}
\usage{
set.nl.variable(
  nl,
  varname,
  values = NULL,
  min = NULL,
  max = NULL,
  step = NULL,
  qfun = NULL
)
}
\arguments{
\item{nl}{nl object}

\item{varname}{name of the variable that should be added}

\item{values}{a values vector, not needed if min, max and step or qfun are defined instead}

\item{min}{minimum value of variable}

\item{max}{maximum value of variable}

\item{step}{step interval between values}

\item{qfun}{distribution function}
}
\value{
nl object
}
\description{
Set a variable in the exepriment of the provided nl object
}
\details{
Set a variable in the exepriment of the provided nl object
}
\examples{
\dontrun{
set.nl.variable(nl, varname="testvar", min=1, max=5, step=0.1, qfun="qunif")
}

}
