% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.rescale.vector.R
\name{util.rescale.vector}
\alias{util.rescale.vector}
\title{Scale a vector of values to a new min and max range (defaults to 0-1 range)}
\usage{
util.rescale.vector(values, new.min = 0, new.max = 1)
}
\arguments{
\item{values}{vector of values to scale}

\item{new.min}{new minimum to scale values to}

\item{new.max}{new maximum to scale values to}
}
\value{
vector of re-scaled values
}
\description{
Scale a vector of values to a new min and max range (defaults to 0-1 range)
}
\details{
Scale a vector of values to a new min and max range (defaults to 0-1 range)
}
\examples{
\dontrun{
tibble::tibble(a=c(3, 2523, 223, 342, 33, 26, 85, 75)) \%>\%
dplyr::mutate(b = util.rescale.vector(a))
}

}
