% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.abm.timeseries.R
\name{doplot.abm.timeseries}
\alias{doplot.abm.timeseries}
\title{Plot EFForTS-ABM time series metrics}
\usage{
doplot.abm.timeseries(nl, metrics = nl@experiment@metrics)
}
\arguments{
\item{nl}{an nl object with attached simulation results containing time series data}

\item{metrics}{a vector of metrics which are plotted as a facet plot}
}
\value{
A ggplot object
}
\description{
Plot EFForTS-ABM time series metrics
}
\details{
Plot EFForTS-ABM time series metrics
}
\examples{
\dontrun{
doplot.abm.timeseries
}

}
