% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.defaults.R
\name{get.abm.defaults}
\alias{get.abm.defaults}
\title{Report default parameterisation:}
\usage{
get.abm.defaults()
}
\value{
A list with EFForTS-ABM default parameters
}
\description{
Report default parameterisation:
}
\details{
Report default parameterisation
}
\examples{
get.abm.defaults()

}
