% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.main.default.R
\name{get.abm.main.default}
\alias{get.abm.main.default}
\title{Get a skeleton for the main parameter file of EFForTS-ABM}
\usage{
get.abm.main.default(crop = NULL)
}
\arguments{
\item{crop}{optionally, select a crop type for defaults}
}
\value{
a list with the parameter format of the main parameter file
}
\description{
Get a skeleton for the main parameter file of EFForTS-ABM
}
\details{
Get a skeleton for the main parameter file of EFForTS-ABM
}
\examples{
get.abm.main.default(crop="rubber")

}
