% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.maxs.R
\name{get.abm.maxs}
\alias{get.abm.maxs}
\title{Report maximum parameterisation:}
\usage{
get.abm.maxs()
}
\value{
A list with EFForTS-ABM maximum parameters
}
\description{
Report maximum parameterisation:
}
\details{
Report maximum parameterisation
}
\examples{
get.abm.maxs()

}
