% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.metrics.patches.R
\name{get.abm.metrics.patches}
\alias{get.abm.metrics.patches}
\title{Report patches metrics as a list}
\usage{
get.abm.metrics.patches()
}
\value{
patches metrics as a list
}
\description{
Report patches metrics as a list
}
\details{
Report patches metrics as a list
}
\examples{
get.abm.metrics.patches()

}
