% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.abm.mins.R
\name{get.abm.mins}
\alias{get.abm.mins}
\title{Report minimum parameterisation:}
\usage{
get.abm.mins()
}
\value{
A list with EFForTS-ABM minimum parameters
}
\description{
Report minimum parameterisation:
}
\details{
Report minimum parameterisation
}
\examples{
get.abm.mins()

}
