% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.del.R
\name{hpc.del}
\alias{hpc.del}
\title{Deletes all files in the specified folder on the remote host (HPC) via ssh}
\usage{
hpc.del(folder = NA, user = NA, host = "transfer.gwdg.de", key = NA)
}
\arguments{
\item{folder}{folder where files are deletd on the remote host}

\item{user}{user name on the remote host}

\item{host}{host server address}

\item{key}{link to the ssh key file on the local system}
}
\description{
Deletes all files in the specified folder on the remote host (HPC) via ssh
}
\details{
Deletes all files in the specified folder on the remote host (HPC) via ssh
}
