% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.download.R
\name{hpc.download}
\alias{hpc.download}
\title{Download files or a complete folder from the remote host (HPC) via ssh}
\usage{
hpc.download(
  from = NA,
  to = NA,
  user = NA,
  host = "transfer.gwdg.de",
  key = NA
)
}
\arguments{
\item{from}{File or folder to copy from the remote host}

\item{to}{destination folder on the local system}

\item{user}{user name on the remote host}

\item{host}{host server address}

\item{key}{link to the ssh key file on the local system}
}
\description{
Sends a file or a complete folder to the remote host (HPC) via ssh
}
\details{
Sends a file or a complete folder to the remote host (HPC) via ssh
}
