% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.upload.R
\name{hpc.upload}
\alias{hpc.upload}
\title{Sends a file or a complete folder to the remote host (HPC) via ssh}
\usage{
hpc.upload(from = NA, to = NA, user = NA, host = "transfer.gwdg.de", key = NA)
}
\arguments{
\item{from}{File or folder to copy}

\item{to}{destination folder on the remote host}

\item{user}{user name on the remote host}

\item{host}{host server address}

\item{key}{link to the ssh key file on the local system}
}
\description{
Sends a file or a complete folder to the remote host (HPC) via ssh
}
\details{
Sends a file or a complete folder to the remote host (HPC) via ssh
}
