% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.GBM.cor.R
\name{prices.GBM.cor}
\alias{prices.GBM.cor}
\title{Creates a matrix of correlated daily price paths using Geometric Brownian Motion.}
\usage{
prices.GBM.cor(N, S0, mu, sigma, cor.mat)
}
\arguments{
\item{N}{Number of days in the path.}

\item{S0}{The initial price of the asset.}

\item{mu}{Drift or average daily continuously compounded returns.}

\item{sigma}{Volatility or standard deviation of daily continuously compounded returns.}

\item{cor.mat}{The correlation matrix of the daility contiuously compounded returns.}
}
\value{
A matrix of simulated daily price paths of length N having the same number of assets as in the mu and sigma vectors. Note that mu and sigma must have the same dimensions.
}
\description{
Get parameter files from efforts-abm folder
}
\details{
Creates a matrix of correlated daily price paths using Geometric Brownian Motion.
}
