% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices.wb}
\alias{prices.wb}
\title{Historical prices data extracted from World Bank (The Pink Sheet)}
\format{historical prices data for rubber and oilpalm}
\usage{
prices.wb
}
\description{
The dataset contains prices for Rubber and Oilpalm in Dollar per ton.
The dataset contains prices from January 1960 to January 2020.
}
\keyword{datasets}
