% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.abm.parfiles.R
\name{set.abm.parfiles}
\alias{set.abm.parfiles}
\title{Create a new parameter sub folder in the parameter subfolder of EFForTS-ABM}
\usage{
set.abm.parfiles(abmfolder, name, main, management, overwrite)
}
\arguments{
\item{abmfolder}{folder destination of EFForTS-ABM model}

\item{name}{name of the new parameter sub folder}

\item{main}{list with main parameter specifications}

\item{management}{nested list with management parameter specifications}

\item{overwrite}{TRUE/FALSE, if TRUE existing parameter files will be overwritten}
}
\description{
Create a new parameter sub folder in the parameter subfolder of EFForTS-ABM
}
\details{
Create a new parameter sub folder in the parameter subfolder of EFForTS-ABM
}
\examples{
\dontrun{
set.abm.parfiles()
}

}
