% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.nl.constant.R
\name{set.nl.constant}
\alias{set.nl.constant}
\title{Set a constant in the exepriment of the provided nl object}
\usage{
set.nl.constant(nl, constname, value)
}
\arguments{
\item{nl}{nl object}

\item{constname}{name of constant which should be updated}

\item{value}{value for constant}
}
\value{
nl object
}
\description{
Set a constant in the exepriment of the provided nl object
}
\details{
Set a constant in the exepriment of the provided nl object
}
\examples{
\dontrun{
set.nl.constant(nl, varname="testconst", value=1)
}

}
