% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.abm.parfiles.R
\name{util.abm.write.main.txt}
\alias{util.abm.write.main.txt}
\title{Write main parameter file for efforts-abm}
\usage{
util.abm.write.main.txt(path, main)
}
\description{
Write main parameter file for efforts-abm
}
\keyword{internal}
