% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.abm.parfiles.R
\name{util.abm.write.management.txt}
\alias{util.abm.write.management.txt}
\title{Write management parameter file for efforts-abm}
\usage{
util.abm.write.management.txt(path, management)
}
\description{
Write management parameter file for efforts-abm
}
\keyword{internal}
